css-align-3/self-alignment/place-self-shorthand-006.html

<!DOCTYPE html>
<title>CSS Box Alignment: place-self shorthand - 'auto' value</title>
<link rel="author" title="Javier Fernandez" href="mailto:jfernandez@igalia.com" />
<link rel="help" href="http://www.w3.org/TR/css3-align/#place-self-property" />
<meta name="assert" content="Check that place-self accepts 'auto' as second value and expands to 'align-self' and 'justify-self' as 'normal'." />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
    #test {
        align-self: center;
        justify-self: start;
        place-self: auto auto;
    }
</style>
<div id="log"></div>
<div id="test"></div>
<script>
    var div = document.getElementById("test");
    var style = getComputedStyle(div);

    test(function() {
        assert_equals(style.getPropertyValue("place-self"),
            "normal normal", "place-self computed value");
    }, "Check place-self: auto - resolved value");

    test(function() {
        assert_equals(style.getPropertyValue("align-self"),
            "normal", "place-self specified value for align-self");
    }, "Check place-self: auto - align-self resolved value");

    test(function() {
        assert_equals(style.getPropertyValue("justify-self"),
           "normal", "place-self specified value for justify-self");
    }, "Check place-self: auto - justify-self resolved value");
</script>
